Move classes for dealing with and displaying results from DVC runs out of DVC interface file
